New BehaviorSession object
